// src/graphql/types.ts
export type Maybe<T> = T | null;
// ...
export type Assignable = {
  /** A list of Users assigned to this object. */
  assignees: UserConnection;
};
// ...
export type CodeOfConduct = Node & {
  __typename?: "CodeOfConduct";
  /** The body of the Code of Conduct */
  body?: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  /** The key for the Code of Conduct */
  key: Scalars["String"];
  /** The formal name of the Code of Conduct */
  name: Scalars["String"];
  /** The HTTP path for this Code of Conduct */
  resourcePath?: Maybe<Scalars["URI"]>;
  /** The HTTP URL for this Code of Conduct */
  url?: Maybe<Scalars["URI"]>;
};
// ...
export type FollowerConnection = {
  __typename?: "FollowerConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<UserEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<User>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars["Int"];
};
// ...
export type Issue = Node &
  Assignable &
  Closable &
  Comment &
  Updatable &
  UpdatableComment &
  Labelable &
  Lockable &
  Reactable &
  RepositoryNode &
  Subscribable &
  UniformResourceLocatable & {
    __typename?: "Issue";
    /** Reason that the conversation was locked. */
    activeLockReason?: Maybe<LockReason>;
    /** A list of Users assigned to this object. */
    assignees: UserConnection;
    /** The actor who authored the comment. */
    author?: Maybe<Actor>;
    /** Author's association with the subject of the comment. */
    authorAssociation: CommentAuthorAssociation;
    /** Identifies the body of the issue. */
    body: Scalars["String"];
    /** Identifies the body of the issue rendered to HTML. */
    bodyHTML: Scalars["HTML"];
    /** Identifies the body of the issue rendered to text. */
    bodyText: Scalars["String"];
    /** `true` if the object is closed (definition of closed may depend on type) */
    closed: Scalars["Boolean"];
    /** Identifies the date and time when the object was closed. */
    closedAt?: Maybe<Scalars["DateTime"]>;
    /** A list of comments associated with the Issue. */
    comments: IssueCommentConnection;
    /** Identifies the date and time when the object was created. */
    createdAt: Scalars["DateTime"];
    /** Check if this comment was created via an email reply. */
    createdViaEmail: Scalars["Boolean"];
    /** Identifies the primary key from the database. */
    databaseId?: Maybe<Scalars["Int"]>;
    /** The actor who edited the comment. */
    editor?: Maybe<Actor>;
    id: Scalars["ID"];
    /** Check if this comment was edited and includes an edit with the creation data */
    includesCreatedEdit: Scalars["Boolean"];
    /** A list of labels associated with the object. */
    labels?: Maybe<LabelConnection>;
    /** The moment the editor made the last edit */
    lastEditedAt?: Maybe<Scalars["DateTime"]>;
    /** `true` if the object is locked */
    locked: Scalars["Boolean"];
    /** Identifies the milestone associated with the issue. */
    milestone?: Maybe<Milestone>;
    /** Identifies the issue number. */
    number: Scalars["Int"];
    /** A list of Users that are participating in the Issue conversation. */
    participants: UserConnection;
    /** List of project cards associated with this issue. */
    projectCards: ProjectCardConnection;
    /** Identifies when the comment was published at. */
    publishedAt?: Maybe<Scalars["DateTime"]>;
    /** A list of reactions grouped by content left on the subject. */
    reactionGroups?: Maybe<Array<ReactionGroup>>;
    /** A list of Reactions left on the Issue. */
    reactions: ReactionConnection;
    /** The repository associated with this node. */
    repository: Repository;
    /** The HTTP path for this issue */
    resourcePath: Scalars["URI"];
    /** Identifies the state of the issue. */
    state: IssueState;
    /** A list of events, comments, commits, etc. associated with the issue. */
    timeline: IssueTimelineConnection;
    /** A list of events, comments, commits, etc. associated with the issue. */
    timelineItems: IssueTimelineItemsConnection;
    /** Identifies the issue title. */
    title: Scalars["String"];
    /** Identifies the date and time when the object was last updated. */
    updatedAt: Scalars["DateTime"];
    /** The HTTP URL for this issue */
    url: Scalars["URI"];
    /** A list of edits to this content. */
    userContentEdits?: Maybe<UserContentEditConnection>;
    /** Can user react to this subject */
    viewerCanReact: Scalars["Boolean"];
    /** Check if the viewer is able to change their subscription status for the repository. */
    viewerCanSubscribe: Scalars["Boolean"];
    /** Check if the current viewer can update this object. */
    viewerCanUpdate: Scalars["Boolean"];
    /** Reasons why the current viewer can not update this comment. */
    viewerCannotUpdateReasons: Array<CommentCannotUpdateReason>;
    /** Did the viewer author this comment. */
    viewerDidAuthor: Scalars["Boolean"];
    /** Identifies if the viewer is watching, not watching, or ignoring the subscribable entity. */
    viewerSubscription?: Maybe<SubscriptionState>;
  };
// ...
export type Language = Node & {
  __typename?: "Language";
  /** The color defined for the current language. */
  color?: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  /** The name of the current language. */
  name: Scalars["String"];
};
// ...
export type License = Node & {
  __typename?: "License";
  /** The full text of the license */
  body: Scalars["String"];
  /** The conditions set by the license */
  conditions: Array<Maybe<LicenseRule>>;
  /** A human-readable description of the license */
  description?: Maybe<Scalars["String"]>;
  /** Whether the license should be featured */
  featured: Scalars["Boolean"];
  /** Whether the license should be displayed in license pickers */
  hidden: Scalars["Boolean"];
  id: Scalars["ID"];
  /** Instructions on how to implement the license */
  implementation?: Maybe<Scalars["String"]>;
  /** The lowercased SPDX ID of the license */
  key: Scalars["String"];
  /** The limitations set by the license */
  limitations: Array<Maybe<LicenseRule>>;
  /** The license full name specified by <https://spdx.org/licenses> */
  name: Scalars["String"];
  /** Customary short name if applicable (e.g, GPLv3) */
  nickname?: Maybe<Scalars["String"]>;
  /** The permissions set by the license */
  permissions: Array<Maybe<LicenseRule>>;
  /** Whether the license is a pseudo-license placeholder (e.g., other, no-license) */
  pseudoLicense: Scalars["Boolean"];
  /** Short identifier specified by <https://spdx.org/licenses> */
  spdxId?: Maybe<Scalars["String"]>;
  /** URL to the license on <https://choosealicense.com> */
  url?: Maybe<Scalars["URI"]>;
};
// ...
export type Node = {
  /** ID of the object. */
  id: Scalars["ID"];
};
// ...
export type Query = {
  __typename?: "Query";
  /** Look up a code of conduct by its key */
  codeOfConduct?: Maybe<CodeOfConduct>;
  /** Look up a code of conduct by its key */
  codesOfConduct?: Maybe<Array<Maybe<CodeOfConduct>>>;
  /** Look up an open source license by its key */
  license?: Maybe<License>;
  /** Return a list of known open source licenses */
  licenses: Array<Maybe<License>>;
  /** Get alphabetically sorted list of Marketplace categories */
  marketplaceCategories: Array<MarketplaceCategory>;
  /** Look up a Marketplace category by its slug. */
  marketplaceCategory?: Maybe<MarketplaceCategory>;
  /** Look up a single Marketplace listing */
  marketplaceListing?: Maybe<MarketplaceListing>;
  /** Look up Marketplace listings */
  marketplaceListings: MarketplaceListingConnection;
  /** Return information about the GitHub instance */
  meta: GitHubMetadata;
  /** Fetches an object given its ID. */
  node?: Maybe<Node>;
  /** Lookup nodes by a list of IDs. */
  nodes: Array<Maybe<Node>>;
  /** Lookup a organization by login. */
  organization?: Maybe<Organization>;
  /** The client's rate limit information. */
  rateLimit?: Maybe<RateLimit>;
  /** Hack to workaround https://github.com/facebook/relay/issues/112 re-exposing the root query object */
  relay: Query;
  /** Lookup a given repository by the owner and repository name. */
  repository?: Maybe<Repository>;
  /** Lookup a repository owner (ie. either a User or an Organization) by login. */
  repositoryOwner?: Maybe<RepositoryOwner>;
  /** Lookup resource by a URL. */
  resource?: Maybe<UniformResourceLocatable>;
  /** Perform a search across resources. */
  search: SearchResultItemConnection;
  /** GitHub Security Advisories */
  securityAdvisories: SecurityAdvisoryConnection;
  /** Fetch a Security Advisory by its GHSA ID */
  securityAdvisory?: Maybe<SecurityAdvisory>;
  /** Software Vulnerabilities documented by GitHub Security Advisories */
  securityVulnerabilities: SecurityVulnerabilityConnection;
  /** Look up a single Sponsors Listing */
  sponsorsListing?: Maybe<SponsorsListing>;
  /** Look up a topic by name. */
  topic?: Maybe<Topic>;
  /** Lookup a user by login. */
  user?: Maybe<User>;
  /** The currently authenticated user. */
  viewer: User;
};
// ...
export enum SearchType {
  /** Returns results matching issues in repositories. */
  Issue = "ISSUE",
  /** Returns results matching repositories. */
  Repository = "REPOSITORY",
  /** Returns results matching users and organizations on GitHub. */
  User = "USER"
}
// ...
export type StargazerConnection = {
  __typename?: "StargazerConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<StargazerEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<User>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars["Int"];
};
// ...
export type StarredRepositoryConnection = {
  __typename?: "StarredRepositoryConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<StarredRepositoryEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<Repository>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars["Int"];
};
// ...
export type RepositoryFragment = { __typename?: "Repository" } & Pick<
  Repository,
  "id" | "description" | "nameWithOwner" | "updatedAt" | "url"
> & {
    primaryLanguage: Maybe<
      { __typename?: "Language" } & Pick<Language, "id" | "name">
    >;
    stargazers: { __typename?: "StargazerConnection" } & Pick<
      StargazerConnection,
      "totalCount"
    >;
  };

export type SearchResultItemFragment =
  | ({ __typename?: "User" } & UserFragment)
  | ({ __typename?: "Repository" } & RepositoryFragment);

export type SearchResultItemConnectionFragment = {
  __typename?: "SearchResultItemConnection";
} & Pick<SearchResultItemConnection, "userCount" | "repositoryCount"> & {
    pageInfo: { __typename?: "PageInfo" } & PageInfoFragment;
    nodes: Maybe<Array<Maybe<SearchResultItemFragment>>>;
  };

export type UserFragment = { __typename?: "User" } & Pick<
  User,
  "id" | "login" | "name" | "avatarUrl" | "bio" | "email" | "location"
> & {
    starredRepositories: { __typename?: "StarredRepositoryConnection" } & Pick<
      StarredRepositoryConnection,
      "totalCount"
    >;
    followers: { __typename?: "FollowerConnection" } & Pick<
      FollowerConnection,
      "totalCount"
    >;
  };

export type SearchQueryVariables = {
  type: SearchType;
  searchTerm: Scalars["String"];
  first?: Maybe<Scalars["Int"]>;
  after?: Maybe<Scalars["String"]>;
  last?: Maybe<Scalars["Int"]>;
  before?: Maybe<Scalars["String"]>;
};

export type SearchQuery = { __typename?: "Query" } & {
  search: {
    __typename?: "SearchResultItemConnection";
  } & SearchResultItemConnectionFragment;
};
// ...
export const SearchDocument = gql`
  query Search(
    $type: SearchType!
    $searchTerm: String!
    $first: Int
    $after: String
    $last: Int
    $before: String
  ) {
    search(
      first: $first
      after: $after
      last: $last
      before: $before
      query: $searchTerm
      type: $type
    ) {
      ...SearchResultItemConnection
    }
  }
  ${SearchResultItemConnectionFragmentDoc}
`;

@Injectable({
  providedIn: "root"
})
export class SearchGQL extends Apollo.Query<SearchQuery, SearchQueryVariables> {
  document = SearchDocument;
}